API : done

+++ FastAPI Project/app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import UserRead, UserCreate, UserUpdate
-from app.db import Post, create_db, get_session
+from app.db import Post, create_db, get_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -29,6 +29,7 @@ app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix=
 async def upload_file(
         file: UploadFile = File(...),
         caption: str = Form(""),
+        user: User = Depends(curr_active_user),
         session: AsyncSession = Depends(get_session)
 ):
     temp_file_path = None
@@ -48,6 +49,7 @@ async def upload_file(
             )
 
         post = Post(
+            user_id=user.id,
             caption=caption,
             url=upload_result.url,
             file_type="video" if file.content_type.startswith("video/") else "image",
@@ -65,27 +67,34 @@ async def upload_file(
         file.file.close()
 
 @app.get("/feed")
-async def get_feed(session: AsyncSession = Depends(get_session)):
+async def get_feed(session: AsyncSession = Depends(get_session), user: User=Depends(curr_active_user)):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
+
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    users_dict = {u.id: u.email for u in users}
 
     posts_data = []
     for post in posts:
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
                 "caption": post.caption,
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
-                "created_at": post.created_at.isoformat()
+                "created_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": users_dict.get(post.user_id, "Unknown")
             }
         )
 
     return {"posts": posts_data}
 
 @app.delete("/posts/{post_id}")
-async def delete(post_id: str, session: AsyncSession = Depends(get_session)):
+async def delete(post_id: str, session: AsyncSession = Depends(get_session), user: User=Depends(curr_active_user)):
     try:
         post_uuid = uuid.UUID(post_id)
         result = await session.execute(select(Post).where(Post.id == post_uuid))
@@ -93,6 +102,9 @@ async def delete(post_id: str, session: AsyncSession = Depends(get_session)):
 
         if not post:
             raise HTTPException(status_code=404, detail="post not found")
+
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="you don't have permission to delete this post")
 
         await session.delete(post)
         await session.commit()
